+++ derive/src/derive/core/convert/mod.rs
@@ -1,2 +1,4 @@
+#[cfg(feature = "convert")]
 pub(crate) mod as_mut;
+#[cfg(feature = "convert")]
 pub(crate) mod as_ref;