+++ derive/src/lib.rs
@@ -1,5 +1,5 @@
 #![recursion_limit = "256"]
-#![doc(html_root_url = "https://docs.rs/auto_enums_derive/0.5.9")]
+#![doc(html_root_url = "https://docs.rs/auto_enums_derive/0.5.10")]
 #![warn(unsafe_code)]
 #![warn(rust_2018_idioms, unreachable_pub)]
 #![warn(clippy::all, clippy::pedantic)]