+++ test_suite/unstable/tests/test.rs
@@ -59,12 +59,12 @@ mod test_futures {
     }
 }
 
+#[cfg(feature = "unstable")]
 mod enum_derive {
     #![allow(dead_code)]
 
     use auto_enums::enum_derive;
 
-    #[cfg(feature = "unstable")]
     #[test]
     fn unstable() {
         #[enum_derive(