+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 0.5.10 - 2019-08-15
+
 * Updated `proc-macro2`, `syn`, and `quote` to 1.0.
 
 * Updated `derive_utils` to 0.9. This improves the error message.